+++ app/ids-loading-indicator/example.js
@@ -1,3 +1,9 @@
+import IdsLayoutGrid from '../../src/ids-layout-grid/ids-layout-grid';
+import IdsLayoutGridCell from '../../src/ids-layout-grid/ids-layout-grid-cell';
+import IdsTooltip from '../../src/ids-tooltip/ids-tooltip';
+import IdsSpinbox from '../../src/ids-spinbox';
+import './index.scss';
+
 document.addEventListener('DOMContentLoaded', () => {
   const determinateIndicators = document.querySelectorAll('ids-loading-indicator[progress]');