+++ app/ids-popup/example.js
@@ -1,7 +1,7 @@
 // Supporting Components
 import IdsButton from '../../src/ids-button/ids-button';
 
-import './index.scss';
+import './example.scss';
 
 document.addEventListener('DOMContentLoaded', () => {
   const triggerId = '#popup-trigger-btn';